:recycle: mon-pix: split congratulations-certification-banner component Co-authored-by: Steph0 <Steph0@users.noreply.github.com>
