+++ mon-pix/app/components/certification-banners/outdated-complementary-certification-banner.hbs
