+++ mon-pix/app/controllers/authenticated/certifications/join.js
@@ -1,22 +1,10 @@
 import Controller from '@ember/controller';
 import { service } from '@ember/service';
-import { tracked } from '@glimmer/tracking';
 import { action } from '@ember/object';
 
 export default class JoinCertificationController extends Controller {
   @service currentUser;
   @service router;
-
-  @tracked displayCongratulationsBanner = true;
-
-  get showCongratulationsBanner() {
-    return this.displayCongratulationsBanner;
-  }
-
-  @action
-  closeBanner() {
-    this.displayCongratulationsBanner = false;
-  }
 
   @action
   changeStep(certificationCandidateId) {
