+++ mon-pix/tests/integration/components/certification-banners/congratulations-certification-banner_test.js
@@ -1,11 +1,11 @@
 import { module, test } from 'qunit';
 import sinon from 'sinon';
-import setupIntlRenderingTest from '../../helpers/setup-intl-rendering';
+import setupIntlRenderingTest from '../../../helpers/setup-intl-rendering';
 import { render } from '@1024pix/ember-testing-library';
 import { click } from '@ember/test-helpers';
 import { hbs } from 'ember-cli-htmlbars';
 
-module('Integration | Component | Congratulations Certification Banner', function (hooks) {
+module('Integration | Component | Certification Banners | Congratulations Certification Banner', function (hooks) {
   setupIntlRenderingTest(hooks);
 
   test('renders a banner indicating the user certifiability', async function (assert) {
@@ -17,7 +17,7 @@ module('Integration | Component | Congratulations Certification Banner', functio
 
     // when
     const screen = await render(
-      hbs`<CongratulationsCertificationBanner @fullName={{this.fullName}} @closeBanner={{this.closeBanner}} @certificationEligibility={{this.certificationEligibility}}/>`,
+      hbs`<CertificationBanners::CongratulationsCertificationBanner @fullName={{this.fullName}} @closeBanner={{this.closeBanner}} @certificationEligibility={{this.certificationEligibility}}/>`,
     );
 
     // then
@@ -33,7 +33,7 @@ module('Integration | Component | Congratulations Certification Banner', functio
     this.set('certificationEligibility', store.createRecord('is-certifiable', {}));
 
     const screen = await render(
-      hbs`<CongratulationsCertificationBanner @fullName={{this.fullName}} @closeBanner={{this.closeBanner}} @certificationEligibility={{this.certificationEligibility}}/>`,
+      hbs`<CertificationBanners::CongratulationsCertificationBanner @fullName={{this.fullName}} @closeBanner={{this.closeBanner}} @certificationEligibility={{this.certificationEligibility}}/>`,
     );
 
     // when
@@ -49,6 +49,13 @@ module('Integration | Component | Congratulations Certification Banner', functio
       test(`renders the complementary certification eligibility special message and picture`, async function (assert) {
         // given
         const store = this.owner.lookup('service:store');
+        const eligibleComplementaryCertifications = [
+          {
+            label: 'CléA Numérique',
+            imageUrl: 'http://www.image-clea.com',
+            isOutdated: false,
+          },
+        ];
         const certificationEligibility = store.createRecord('is-certifiable', {
           complementaryCertifications: [
             {
@@ -58,13 +65,14 @@ module('Integration | Component | Congratulations Certification Banner', functio
             },
           ],
         });
+        this.set('eligibleComplementaryCertifications', eligibleComplementaryCertifications);
         this.set('certificationEligibility', certificationEligibility);
         this.set('fullName', 'Fifi Brindacier');
         this.set('closeBanner', () => {});
 
         // when
         const screen = await render(
-          hbs`<CongratulationsCertificationBanner @certificationEligibility={{this.certificationEligibility}} @fullName={{this.fullName}} @closeBanner={{this.closeBanner}}/>`,
+          hbs`<CertificationBanners::CongratulationsCertificationBanner @eligibleComplementaryCertifications={{this.eligibleComplementaryCertifications}} @complementaryCertifications={{this.certificationEligibility.complementaryCertifications}} @fullName={{this.fullName}} @closeBanner={{this.closeBanner}}/>`,
         );
 
         // then
@@ -78,19 +86,28 @@ module('Integration | Component | Congratulations Certification Banner', functio
       test(`renders the multiple complementary certification eligibility pluralized message and pictures`, async function (assert) {
         // given
         const store = this.owner.lookup('service:store');
+        const eligibleComplementaryCertifications = [
+          {
+            label: 'CléA Numérique',
+            imageUrl: 'http://www.image-clea.com',
+            isOutdated: false,
+          },
+          { label: 'Pix+ Édu 1er degré Confirmé', imageUrl: 'http://www.image-clea.com', isOutdated: false },
+        ];
         const certificationEligibility = store.createRecord('is-certifiable', {
           complementaryCertifications: [
             { label: 'CléA Numérique', imageUrl: 'http://www.image-clea.com', isOutdated: false },
             { label: 'Pix+ Édu 1er degré Confirmé', imageUrl: 'http://www.image-clea.com', isOutdated: false },
           ],
         });
+        this.set('eligibleComplementaryCertifications', eligibleComplementaryCertifications);
         this.set('certificationEligibility', certificationEligibility);
         this.set('fullName', 'Fifi Brindacier');
         this.set('closeBanner', () => {});
 
         // when
         const screen = await render(
-          hbs`<CongratulationsCertificationBanner @certificationEligibility={{this.certificationEligibility}} @fullName={{this.fullName}} @closeBanner={{this.closeBanner}}/>`,
+          hbs`<CertificationBanners::CongratulationsCertificationBanner @eligibleComplementaryCertifications={{this.eligibleComplementaryCertifications}} @complementaryCertifications={{this.certificationEligibility.complementaryCertifications}} @fullName={{this.fullName}} @closeBanner={{this.closeBanner}}/>`,
         );
 
         // then
@@ -99,55 +116,6 @@ module('Integration | Component | Congratulations Certification Banner', functio
         assert.ok(screen.getByText('Pix+ Édu 1er degré Confirmé'));
         assert.ok(screen.getByRole('img', { name: 'CléA Numérique' }));
         assert.ok(screen.getByRole('img', { name: 'Pix+ Édu 1er degré Confirmé' }));
-      });
-    });
-
-    module('When there are outdated complementary certifications', function () {
-      test(`renders the outdated complementary certification`, async function (assert) {
-        // given
-        const store = this.owner.lookup('service:store');
-        const certificationEligibility = store.createRecord('is-certifiable', {
-          complementaryCertifications: [
-            { label: 'CléA Numérique', imageUrl: 'http://www.image-clea.com', isOutdated: true },
-            { label: 'Pix+ Édu 1er degré Confirmé', imageUrl: 'http://www.image-clea.com', isOutdated: true },
-          ],
-        });
-        this.set('certificationEligibility', certificationEligibility);
-        this.set('closeBanner', () => {});
-
-        // when
-        const screen = await render(
-          hbs`<CongratulationsCertificationBanner @certificationEligibility={{this.certificationEligibility}} @fullName={{this.fullName}} @closeBanner={{this.closeBanner}}/>`,
-        );
-
-        // then
-        assert.dom(screen.getByRole('img', { name: 'CléA Numérique' })).exists();
-        assert
-          .dom(screen.getByText("Vous n'êtes plus éligible à la certification CléA Numérique suite à son évolution."))
-          .exists();
-        assert
-          .dom(
-            screen.getAllByText(
-              "Recontactez votre établissement ou l’organisation vous ayant proposé le parcours afin de repasser une campagne et ainsi redevenir éligible. Votre progression a été conservée et vous n'aurez qu'à jouer les nouvelles épreuves, cela devrait être rapide.",
-            )[0],
-          )
-          .exists();
-
-        assert.ok(screen.getByRole('img', { name: 'Pix+ Édu 1er degré Confirmé' }));
-        assert
-          .dom(
-            screen.getByText(
-              "Vous n'êtes plus éligible à la certification Pix+ Édu 1er degré Confirmé suite à son évolution.",
-            ),
-          )
-          .exists();
-        assert
-          .dom(
-            screen.getAllByText(
-              "Recontactez votre établissement ou l’organisation vous ayant proposé le parcours afin de repasser une campagne et ainsi redevenir éligible. Votre progression a été conservée et vous n'aurez qu'à jouer les nouvelles épreuves, cela devrait être rapide.",
-            )[1],
-          )
-          .exists();
       });
     });
   });
